+++ integration-tests/tests/src/dynamic-schema-updates.ts
@@ -18,11 +18,11 @@
 
 import { expect } from "chai";
 
-import { PersonAndDogSchema } from "./schemas/person-and-dogs";
+import { DogSchema, PersonSchema } from "./schemas/person-and-dogs";
 
 describe("realm._updateSchema", () => {
     it("is a function", () => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         expect(realm.schema).to.be.an("array");
         // There is a function defined on the Realm
         expect(realm._updateSchema).to.be.a("function");
@@ -33,7 +33,7 @@ describe("realm._updateSchema", () => {
     });
 
     it("creates a class schema from a name", () => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         realm.write(() => {
             realm._updateSchema([
                 ...realm.schema,
@@ -45,7 +45,7 @@ describe("realm._updateSchema", () => {
     });
 
     it("creates a class schema from a name and properties", () => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         realm.write(() => {
             realm._updateSchema([
                 ...realm.schema,
@@ -68,7 +68,7 @@ describe("realm._updateSchema", () => {
     });
 
     it("creates a property on an existing class", () => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         // Copy the schema
         const updatedSchema = [...realm.schema];
         // Locate the Dog schema
@@ -119,7 +119,7 @@ describe("realm._updateSchema", () => {
     });
 
     it("can use a newly added class", () => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         realm.write(() => {
             realm._updateSchema([
                 ...realm.schema,
@@ -133,7 +133,7 @@ describe("realm._updateSchema", () => {
     });
 
     it("fires the schema change event", done => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         realm.addListener("schema", () => {
             expect(realm.schema).to.have.length(3);
             const objectSchemaNames = realm.schema.map(s => s.name);
@@ -150,7 +150,7 @@ describe("realm._updateSchema", () => {
     });
 
     it("throws if creating a class schema outside of a transaction", () => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         expect(() => {
             realm._updateSchema([
                 ...realm.schema,
@@ -160,7 +160,7 @@ describe("realm._updateSchema", () => {
     });
 
     it("throws if asked to create a class that already exists", () => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         expect(() => {
             realm.write(() => {
                 realm._updateSchema([
@@ -172,7 +172,7 @@ describe("realm._updateSchema", () => {
     });
 
     it("throws if called without a schema object", () => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         expect(() => {
             realm.write(() => {
                 (realm as any)._updateSchema();
@@ -181,7 +181,7 @@ describe("realm._updateSchema", () => {
     });
 
     it("throws if called with an unexpected type", () => {
-        const realm = new Realm({ schema: PersonAndDogSchema });
+        const realm = new Realm({ schema: [PersonSchema, DogSchema] });
         expect(() => {
             realm.write(() => {
                 (realm as any)._updateSchema("w00t");