+++ lib/index.d.ts
@@ -192,6 +192,12 @@ declare namespace Realm {
     }
 
     /**
+     * RealmJsonSerializeReplacer solves circular structures when serializing Realm entities
+     * @example JSON.stringify(realm.objects("Person"), Realm.RealmJsonSerializeReplacer)
+     */
+    const JsonSerializationReplacer: (key: string, val: any) => any;
+
+    /**
      * SortDescriptor
      * @see { @link https://realm.io/docs/javascript/latest/api/Realm.Collection.html#~SortDescriptor }
      */
@@ -204,6 +210,11 @@ declare namespace Realm {
     interface Collection<T> extends ReadonlyArray<T> {
         readonly type: PropertyType;
         readonly optional: boolean;
+
+        /**
+         * @returns An object for JSON serialization.
+         */
+        toJSON(): Array<any>;
 
         description(): string;
 
@@ -813,6 +824,32 @@ interface NamedSubscription {
     timeToLive: number;
 }
 
+// Remove & exchange for Omit if TypeScript is bumped beyond v3.5
+type Omitter<T, K extends keyof any> = Pick<T, Exclude<keyof T, K>>;
+
+/**
+ * Extracts an intersection of keys from T, where the value extends the given PropType.
+ */
+type ExtractPropertyNamesOfType<T, PropType> = {
+    [K in keyof T]: T[K] extends PropType ? K : never
+}[keyof T];
+
+/**
+ * Exchanges properties defined as Realm.List<Model> with an optional Array<Model | RealmInsertionModel<Model>>.
+ */
+type RealmOptionalListMappedModel<T> = {
+    [K in keyof T]?: T[K] extends Realm.List<infer GT> ? Array<GT | RealmInsertionModel<GT>> : never
+}
+
+/**
+ * Joins T stripped of all keys which value extends Realm.Collection and all inherited from Realm.Object,
+ * with only the keys which value extends Realm.List, remapped as Arrays.
+ */
+//
+type RealmInsertionModel<T> =
+    Omitter<Omitter<T, keyof Realm.Object>, ExtractPropertyNamesOfType<T, Realm.Collection<any>>>
+    & RealmOptionalListMappedModel<Pick<T, ExtractPropertyNamesOfType<T, Realm.List<any>>>>
+
 declare class Realm {
     static defaultPath: string;
 
@@ -899,22 +936,40 @@ declare class Realm {
     close(): void;
 
     /**
-     * @param  {string|Realm.ObjectClass|Function} type
-     * @param  {T&Realm.ObjectPropsType} properties
+     * @param  {string} type
+     * @param  {T} properties
+     * @param  {boolean} update?
+     * @returns T & Realm.Object
+     *
+     * @deprecated, to be removed in future versions. Use `create(type, properties, UpdateMode)` instead.
+     */
+    create<T>(type: string, properties: RealmInsertionModel<T>, update?: boolean): T & Realm.Object
+
+    /**
+     * @param  {Class} type
+     * @param  {T} properties
      * @param  {boolean} update?
      * @returns T
      *
      * @deprecated, to be removed in future versions. Use `create(type, properties, UpdateMode)` instead.
      */
-    create<T>(type: string | Realm.ObjectClass | Function, properties: T | Realm.ObjectPropsType, update?: boolean): T;
+    create<T extends Realm.Object>(type: {new(...arg: any[]): T; }, properties: RealmInsertionModel<T>, update?: boolean): T
 
     /**
-     * @param  {string|Realm.ObjectClass|Function} type
-     * @param  {T&Realm.ObjectPropsType} properties
+     * @param  {string} type
+     * @param  {T} properties
+     * @param  {Realm.UpdateMode} mode? If not provided, `Realm.UpdateMode.Never` is used.
+     * @returns T & Realm.Object
+     */
+    create<T>(type: string, properties: RealmInsertionModel<T>, mode?: Realm.UpdateMode): T & Realm.Object
+
+    /**
+     * @param  {Class} type
+     * @param  {T} properties
      * @param  {Realm.UpdateMode} mode? If not provided, `Realm.UpdateMode.Never` is used.
      * @returns T
      */
-    create<T>(type: string | Realm.ObjectClass | Function, properties: T | Realm.ObjectPropsType, mode?: Realm.UpdateMode): T;
+    create<T extends Realm.Object>(type: {new(...arg: any[]): T; }, properties: RealmInsertionModel<T>, mode?: Realm.UpdateMode): T
 
     /**
      * @param  {Realm.Object|Realm.Object[]|Realm.List<any>|Realm.Results<any>|any} object
@@ -947,10 +1002,16 @@ declare class Realm {
     objectForPrimaryKey<T>(type: string | Realm.ObjectType | Function, id: string): T & Realm.Object | undefined;
 
     /**
-     * @param  {string|Realm.ObjectType|Function} type
-     * @returns Realm
+     * @param  {string} type
+     * @returns Realm.Results<T & Realm.Object>
      */
-    objects<T>(type: string | Realm.ObjectType | Function): Realm.Results<T & Realm.Object>;
+    objects<T>(type: string): Realm.Results<T & Realm.Object>;
+
+    /**
+     * @param  {Class} type
+     * @returns Realm.Results<T>
+     */
+    objects<T extends Realm.Object>(type: {new(...arg: any[]): T; }): Realm.Results<T>;
 
     /**
      * @param  {string} name